Change array size variables from int to size_t

+++ src/data/BloomFilter.h
@@ -8,7 +8,7 @@ private:
     BloomStorage& storage;
     std::vector<int> bitArray;
 public:
-    BloomFilter(BloomStorage& storage, int arraySize);
+    BloomFilter(BloomStorage& storage, size_t arraySize);
     bool contain(const std::vector<int>& hashResults) const;
     void add(const std::vector<int>& hashResults);
 };
